requirements.txt
openai
python-dotenv
GitPython
cryptography